pyproject.toml
[tool.poetry]
name = "learnfastapi"
version = "0.1.0"
description = ""
authors = ["kouhkan <amirkouhkan1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

